worker/python/requirements.txt
# Core runtime
pandas==2.2.2
numpy==1.26.4
scipy==1.11.4
scikit-learn==1.4.2
lightgbm==4.3.0
joblib==1.4.2
yfinance==0.2.38
requests==2.32.3
matplotlib==3.8.4
optuna==3.6.1
tqdm==4.66.4
pydantic==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.2
ccxt==4.3.77
# Technical indicators helper (optional; we provide fallbacks)
ta==0.11.0

# Modern UI Framework
PySide6==6.6.0
pyqtgraph==0.13.3

# Dev & QA
yapf==0.40.2
black==24.4.2
flake8==7.0.0
pytest==8.2.2
pytest-cov==5.0.0
pyinstaller==6.10.0

